Update value of a linear predictor as function of a single coordinate change

Function for updating the linear predictor with n actions rather than nn_vars actions if naively doing a matrix-vector product of X %% beta

Usage

update_linear_predictor(new_beta_j, current_beta_j, current_eta, X_j)

Arguments

new_beta_j

a numeric new value of the j'th component of the beta parameter vector

current_beta_j

the numeric current value of the j'th component of the beta parameter vector

current_eta

the current value of the linear predictor corresponding to the current_beta value

X_j

the j'th column in the design matrix